+++ src/main/java/library/BookRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
     List<Book> findAll();
     Book findByTitleAndAuthor(String title, String author);
 
-    Optional<Book> findById(Long bookId);
+    Optional<Book> findById(Integer bookId);
 }
